requirements.txt
numpy>=1.20.0
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
opencv-python>=4.5.0
scipy>=1.7.0
scikit-image>=0.18.0
matplotlib>=3.4.0